+++ studio/src/app/helpers/editor/image.helper.ts
@@ -116,7 +116,7 @@ export class ImageHelper {
 
           this.didChange.emit(selectedElement);
 
-          await (selectedElement as any).loadBackground();
+          await (selectedElement as HTMLDeckgoDeckElement).loadBackground();
         } else if (selectedElement.hasAttribute('custom-background')) {
           selectedElement.removeChild(currentSlotElement);
 
@@ -126,7 +126,7 @@ export class ImageHelper {
 
           // Refresh background, deck might have one defined which need to be replicated on the slide which
           if (selectedElement.parentElement) {
-            await (selectedElement.parentElement as any).loadBackground();
+            await (selectedElement.parentElement as HTMLDeckgoDeckElement).loadBackground();
           }
         }
       }
@@ -143,12 +143,12 @@ export class ImageHelper {
     const deckgImg: DeckgoImgAction | undefined = ImageActionUtils.extractAttributes(image);
 
     if (deckgImg !== undefined) {
-      (img as any).imgSrc = deckgImg.src;
-      (img as any).imgAlt = deckgImg.label;
+      (img as HTMLDeckgoLazyImgElement).imgSrc = deckgImg.src;
+      (img as HTMLDeckgoLazyImgElement).imgAlt = deckgImg.label;
     }
 
     if (image && image !== undefined && image.hasOwnProperty('downloadUrl')) {
-      (img as any).customLoader = true;
+      (img as HTMLDeckgoLazyImgElement).customLoader = true;
 
       // We have to add the information as attributes because slots are going to be cloned to the slides background
       if (background) {
@@ -231,7 +231,7 @@ export class ImageHelper {
     selectedElement.appendChild(div);
 
     if (deck) {
-      await (selectedElement as any).loadBackground();
+      await (selectedElement as HTMLDeckgoDeckElement).loadBackground();
     } else {
       selectedElement.setAttribute('custom-background', '');
     }
@@ -255,7 +255,7 @@ export class ImageHelper {
     if (deck) {
       // prettier-ignore
       img.addEventListener("lazyImgDidLoad", async () => {
-        await (selectedElement as any).loadBackground();
+        await (selectedElement as HTMLDeckgoDeckElement).loadBackground();
         this.didChange.emit(selectedElement);
       }, { once: true });
     } else {